+++ src/ts/Native/NativeApi.ts
@@ -4,14 +4,16 @@ import { Platform } from 'Constants/Platform';
 export enum ApiPackage {
     CORE,
     ADS,
-    AR
+    AR,
+    BANNER
 }
 
 export abstract class NativeApi {
     private static _apiPackageMapping = {
         [ApiPackage.CORE]: {android: 'com.unity3d.services.core.api', ios: 'USRVApi'},
         [ApiPackage.ADS]: {android: 'com.unity3d.services.ads.api', ios: 'UADSApi'},
-        [ApiPackage.AR]: {android: 'com.unity3d.services.ar.api', ios: 'UARApi'}
+        [ApiPackage.AR]: {android: 'com.unity3d.services.ar.api', ios: 'UARApi'},
+        [ApiPackage.BANNER]: {android: 'com.unity3d.services.banner.api', ios: 'UADSApi'}
     };
 
     protected _nativeBridge: NativeBridge;