+++ src/ts/Parsers/CometCampaignParser.ts
@@ -13,6 +13,23 @@ import { AdUnitStyle } from 'Models/AdUnitStyle';
 import { CustomFeatures } from 'Utilities/CustomFeatures';
 import { Diagnostics } from 'Utilities/Diagnostics';
 
+// Events marked with // are currently sent, but are unused - waiting for BI to confirm if they want them sent
+export enum ICometTrackingUrlEvents {
+    IMPRESSION = 'EventImpression', //
+    START = 'EventStart',
+    CLICK = 'EventClick',
+    ENDCARD_CLICK = 'EventVideoEndCardClick', //
+    FIRST_QUARTILE = 'EventFirstQuartile',
+    MIDPOINT = 'EventMidpoint',
+    THIRD_QUARTILE = 'EventThirdQuartile',
+    ERROR = 'EventError',
+    STALLED = 'EventStalled', //
+    LOADED = 'EventLoaded',
+    SHOW = 'EventShow', //
+    COMPLETE = 'EventComplete',
+    SKIP = 'EventSkip'
+}
+
 export class CometCampaignParser extends CampaignParser {
     public static ContentType = 'comet/campaign';
 
@@ -73,7 +90,7 @@ export class CometCampaignParser extends CampaignParser {
 
             const mraidCampaign = new MRAIDCampaign(parameters);
 
-            if(CustomFeatures.isPlayableConfigurationEnabled(json.mraidUrl)) {
+            if (CustomFeatures.isPlayableConfigurationEnabled(json.mraidUrl)) {
                 const playableConfigurationUrl = json.mraidUrl.replace(/index\.html/, 'configuration.json');
                 request.get(playableConfigurationUrl).then(configurationResponse => {
                     try {
@@ -106,7 +123,8 @@ export class CometCampaignParser extends CampaignParser {
                 videoEventUrls: this.validateAndEncodeVideoEventUrls(json.videoEventUrls, session),
                 bypassAppSheet: json.bypassAppSheet,
                 store: storeName,
-                adUnitStyle: this.parseAdUnitStyle(json.adUnitStyle)
+                adUnitStyle: this.parseAdUnitStyle(json.adUnitStyle),
+                trackingUrls: response.getTrackingUrls()
             };
 
             if(json.trailerDownloadable && json.trailerDownloadableSize && json.trailerStreaming) {